fix: messages generation not conforming to rename
